Pull ArchivalMediaBagParser out into its own model
Required a change to the way the BarcodeComponentDict worked
