+++ .rubocop.yml
@@ -63,9 +63,6 @@ Metrics/CyclomaticComplexity:
   Exclude:
     - 'app/models/concerns/linked_data/linked_resource_factory.rb'
     - 'app/services/file_appender.rb'
-Metrics/LineLength:
-  Exclude:
-    - 'spec/jobs/ingest_archival_media_bag_job_spec.rb'
 Metrics/MethodLength:
   Exclude:
     - 'db/migrate/**/*'
